backend/requirements.txt
# eldntr/ai-rextra/requirements.txt

Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
langgraph==0.0.65
openpyxl==3.1.2
google-generativeai==0.3.2
pymongo==4.8.0
flasgger==0.9.7.1